clean up Tracer
